feat: ajouter des restrictions d'accès basées sur le rôle utilisateur

+++ src/controllers/animal.controller.ts
@@ -3,6 +3,11 @@ import * as service from '../services/animal.service';
 
 export const getAll = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
     const page = Number(req.query.page) || 1;
     const limit = Number(req.query.limit) || 10;
     const search = req.query.search as string;
@@ -16,8 +21,14 @@ export const getAll = async (req: Request, res: Response, next: NextFunction) =>
 
 export const getOne = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
     const id = Number(req.params.id);
     const animal = await service.getById(id);
+
+    if (role !== 'VETERINAIRE' && req.user?.id !== animal?.ownerId) {
+      res.status(401).json({ error: "Unauthorized" });
+      return;
+    }
     
     if (!animal) {
       res.status(404).json({ error: "Animal non trouvé" });
@@ -40,6 +51,11 @@ export const create = async (req: Request, res: Response, next: NextFunction) =>
 
 export const update = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
     const id = Number(req.params.id);
     const updatedAnimal = await service.update(id, req.body);
     res.json(updatedAnimal);
@@ -51,6 +67,12 @@ export const update = async (req: Request, res: Response, next: NextFunction) =>
 // DELETE
 export const remove = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
+    
     const id = Number(req.params.id);
     await service.remove(id);
     res.status(204).send();
