+++ src/controllers/vaccine.controller.ts
@@ -6,7 +6,12 @@ export const getAll = async (req: Request, res: Response, next: NextFunction) =>
     const page = Number(req.query.page) || 1;
     const limit = Number(req.query.limit) || 10;
     const search = req.query.search as string;
-    
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
+
     const result = await service.getAll(page, limit, search);
     res.json(result);
   } catch (error) {
@@ -19,6 +24,7 @@ export const getOne = async (req: Request, res: Response, next: NextFunction) =>
     const id = Number(req.params.id);
     const vaccine = await service.getById(id);
     
+
     if (!vaccine) {
       res.status(404).json({ error: "Vaccin non trouvé" });
       return;
@@ -31,6 +37,11 @@ export const getOne = async (req: Request, res: Response, next: NextFunction) =>
 
 export const create = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
     const newVaccine = await service.create(req.body);
     res.status(201).json(newVaccine);
   } catch (error) {
@@ -40,6 +51,11 @@ export const create = async (req: Request, res: Response, next: NextFunction) =>
 
 export const update = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
     const id = Number(req.params.id);
     const updatedVaccine = await service.update(id, req.body);
     res.json(updatedVaccine);
@@ -50,6 +66,11 @@ export const update = async (req: Request, res: Response, next: NextFunction) =>
 
 export const remove = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    const role = req.user?.role;
+    if (role !== 'VETERINAIRE') {
+      res.status(401);
+      return;
+    }
     const id = Number(req.params.id);
     await service.remove(id);
     res.status(204).send();
